+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 3.1.20 / 2017-11-28
 
 * `fix`: CitySelectField readOnly fails to work
+* `fix`: `SrollList` only load one page after refresh
+* `refactor`: rename InfiniteScroll's getRef to getDOMNode
 
 ## 3.1.19 / 2017-11-28